Homework 3: Add 3-layered architecture, Add routing
